+++ src/window/help.hpp
@@ -39,8 +39,8 @@ namespace Ui
       void Redraw();
       void Print(uint32_t line) const;
       void Confirm();
-      void Scroll(int32_t scrollCount);
-      void ScrollTo(uint16_t scrollLine);
+      void Scroll(int64_t scrollCount);
+      void ScrollTo(uint32_t scrollLine);
 
       std::string SearchPattern(int32_t id) const { return help_.Get(id); }
