+++ src/window/scrollwindow.hpp
@@ -47,8 +47,8 @@ namespace Ui
    public:
       virtual void Print(uint32_t line) const;
       virtual void Resize(int rows, int columns);
-      virtual void Scroll(int32_t scrollCount);
-      virtual void ScrollTo(uint16_t scrollLine);
+      virtual void Scroll(int64_t scrollCount);
+      virtual void ScrollTo(uint32_t scrollLine);
       virtual void ScrollToFirstMatch(std::string const & input) { }
       virtual void ScrollToCurrent() { }
       virtual void ResetSelection() {}
@@ -70,7 +70,7 @@ namespace Ui
       virtual void Save(std::string const & name) {}
 
    public:
-      virtual size_t BufferSize() const { return WindowBuffer().Size(); }
+      virtual uint32_t BufferSize() const { return WindowBuffer().Size(); }
 
    public:
       std::string const & Name() const;
@@ -84,12 +84,12 @@ namespace Ui
       virtual bool IsSelected(uint32_t line) const { return false; }
       uint32_t FirstLine()   const;
       uint32_t LastLine()    const { return (BufferSize() < ScrollLine()) ? BufferSize() : ScrollLine(); }
-      virtual  uint16_t CurrentLine() const { return FirstLine(); }
+      virtual  uint32_t CurrentLine() const { return FirstLine(); }
 
    protected:
       void ResetScroll();
-      void SetScrollLine(uint16_t scrollLine);
-      uint16_t ScrollLine() const;
+      void SetScrollLine(uint32_t scrollLine);
+      uint32_t ScrollLine() const;
 
    protected:
       void SoftRedrawOnSetting(Setting::ToggleSettings setting);
@@ -106,7 +106,7 @@ namespace Ui
       Main::Settings &           settings_;
       Ui::Screen &               screen_;
       std::string                name_;
-      uint16_t                   scrollLine_;
+      uint32_t                   scrollLine_;
       bool                       autoScroll_;
    };
 }
