+++ src/window/selectwindow.cpp
@@ -82,7 +82,7 @@ void SelectWindow::Resize(int rows, int columns)
    LimitCurrentSelection();
 }
 
-void SelectWindow::Scroll(int32_t scrollCount)
+void SelectWindow::Scroll(int64_t scrollCount)
 {
    currentLine_ += scrollCount;
    LimitCurrentSelection();
@@ -93,7 +93,7 @@ void SelectWindow::Scroll(int32_t scrollCount)
    }
 }
 
-void SelectWindow::ScrollTo(uint16_t scrollLine)
+void SelectWindow::ScrollTo(uint32_t scrollLine)
 {
    int64_t oldSelection = currentLine_;
    currentLine_    = (static_cast<int64_t>(scrollLine));
@@ -117,7 +117,7 @@ void SelectWindow::ScrollTo(uint16_t scrollLine)
    }
 }
 
-uint16_t SelectWindow::CurrentLine() const
+uint32_t SelectWindow::CurrentLine() const
 {
    return currentLine_;
 }
@@ -196,7 +196,7 @@ void SelectWindow::LimitCurrentSelection()
    {
       currentLine_ = 0;
    }
-   else if ((currentLine_ >= static_cast<int32_t>(BufferSize())) && (BufferSize() > 0))
+   else if ((currentLine_ >= static_cast<int64_t>(BufferSize())) && (BufferSize() > 0))
    {
       currentLine_ = BufferSize() - 1;
    }
